First Adam-revision of the Manifesto
Few changes; firstly, remove 'religious' as that would be deemed as
processing personal data.

Secondly, ports. Ditch in-secure protocols, the
should-have-been-made-obsolete POP, whack in a few others. Sort by
protocol number (per IANA).

Few other changes on food/accom/internets, too.

Still a work in progress, but things to do, and rather than just
bitch, be proactive.

Doing in a personal capacity (per repo), but may be able to do the
company one, in which case we'll update the supporters to accurately
reflect the company make-up as it currently stands.