Adding section on 'attitude to winners'
I have heard a number of developers say that having participated in, and won, a hackday, sponsors have felt that they are some kind of prize horse they can show off for marketing purposes. This is a poor idea and likely to make shy or nervous participants avoid future events.